requirements.txt
discord.py
PyNaCl
yt-dlp
requests
psutil
colorama
gTTS
langdetect
pillow
tflink
deep-translator